Fix warnings
Former-commit-id: 3e3f473a072195a86b1bd8da69d32e0fb5d6e70c

+++ ZP/ZPPOrderTimeChooserVC.m
@@ -117,7 +117,7 @@ static NSString *ZPPNoInternetConnectionVCIdentifier = @"ZPPNoInternetConnection
         balanceLabel.font = font;
         balanceLabel.textAlignment = NSTextAlignmentCenter;
         balanceLabel.text =
-                    [NSString stringWithFormat:@"Бонусы: %ld%@", balance, ZPPRoubleSymbol];
+                    [NSString stringWithFormat:@"Бонусы: %ld%@", (long)balance, ZPPRoubleSymbol];
         stackHeight += font.lineHeight + stackSpacing;
         [self.totalPriceDetailsStackView addArrangedSubview:balanceLabel];
     }
